changes are the fastest I could create for AVR for Fletcher32 (use of the local variable keeps the build-CI happy)

+++ Fletcher32.h
@@ -20,7 +20,7 @@ public:
   void     add(uint16_t value);
   void     add(const uint16_t * array, uint16_t length);
 
-  uint32_t getFletcher() { return (((uint32_t) _s2) << 16) | ((uint32_t) _s1); };
+  uint32_t getFletcher() { return (((uint32_t)_s2) << 16) | _s1; };
   uint32_t count()       { return _count; };
 
 private:
